modifikasi percobaan 1

+++ Jobsheet 7/ForKelipatan23.java
@@ -5,7 +5,7 @@ public class ForKelipatan23 {
     Scanner input23 = new Scanner(System.in);
 
     int kelipatan, counter=0;
-    double jumlah=0, rerata;
+    float jumlah=0, rerata;
 
     System.out.println("Masukkan bilangan kelipatan (1-9): ");
     kelipatan=input23.nextInt();
@@ -21,8 +21,8 @@ public class ForKelipatan23 {
     rerata = jumlah / counter;
 
     System.out.printf("Banyaknya bilangan %d dari 1 sampai 50 adalah %d\n", kelipatan, counter);
-    System.out.printf("Total bilangan kelipatan %d dari 1 sampai 50 adalah %f\n", kelipatan, jumlah);
-    System.out.printf("Rata-rata dari bilangan-bilangan kelipatan %d adalah %f", kelipatan, rerata);
+    System.out.printf("Total bilangan kelipatan %d dari 1 sampai 50 adalah %.0f\n", kelipatan, jumlah);
+    System.out.printf("Rata-rata dari bilangan-bilangan kelipatan %d adalah %.2f", kelipatan, rerata);
 
   } 
 }
